added FFT, no overlap

+++ src/fft/AudioFFT/AudioFFT.cpp
@@ -59,8 +59,8 @@ namespace audiofft
       AudioFFTImpl& operator=(const AudioFFTImpl&) = delete;
       virtual ~AudioFFTImpl() = default;
       virtual void init(size_t size) = 0;
-      virtual void fft(const float* data, float* re, float* im) = 0;
-      virtual void ifft(float* data, const float* re, const float* im) = 0;
+      virtual void fft(const double* data, double* re, double* im) = 0;
+      virtual void ifft(double* data, const double* re, const double* im) = 0;
     };
 
 
@@ -132,7 +132,7 @@ namespace audiofft
       }
     }
 
-    virtual void fft(const float* data, float* re, float* im) override
+    virtual void fft(const double* data, double* re, double* im) override
     {
       // Convert into the format as required by the Ooura FFT
       detail::ConvertBuffer(_buffer.data(), data, _size);
@@ -143,12 +143,12 @@ namespace audiofft
       {
         double* b = _buffer.data();
         double* bEnd = b + _size;
-        float *r = re;
-        float *i = im;
+        double *r = re;
+        double *i = im;
         while (b != bEnd)
         {
-          *(r++) = static_cast<float>(*(b++));
-          *(i++) = static_cast<float>(-(*(b++)));
+          *(r++) = static_cast<double>(*(b++));
+          *(i++) = static_cast<double>(-(*(b++)));
         }
       }
       const size_t size2 = _size / 2;
@@ -157,14 +157,14 @@ namespace audiofft
       im[size2] = 0.0;
     }
 
-    virtual void ifft(float* data, const float* re, const float* im) override
+    virtual void ifft(double* data, const double* re, const double* im) override
     {
       // Convert into the format as required by the Ooura FFT
       {
         double* b = _buffer.data();
         double* bEnd = b + _size;
-        const float *r = re;
-        const float *i = im;
+        const double *r = re;
+        const double *i = im;
         while (b != bEnd)
         {
           *(b++) = static_cast<double>(*(r++));
@@ -865,7 +865,7 @@ namespace audiofft
       }
     }
 
-    virtual void fft(const float* data, float* re, float* im) override
+    virtual void fft(const double* data, double* re, double* im) override
     {
       size_t complexNumbersCount = _operationalBufferSize / 2;
       ippsFFTFwd_RToCCS_32f(
@@ -885,7 +885,7 @@ namespace audiofft
       }
     }
 
-    virtual void ifft(float* data, const float* re, const float* im) override
+    virtual void ifft(double* data, const double* re, const double* im) override
     {
       size_t complexNumbersCount = _operationalBufferSize / 2;
 
@@ -904,7 +904,7 @@ namespace audiofft
       );
 
       // scaling
-      const float factor = 1.0f / static_cast<float>(_size);
+      const double factor = 1.0f / static_cast<double>(_size);
       ippsMulC_32f_I(factor, data, _size);
     }
 
@@ -987,7 +987,7 @@ namespace audiofft
       }
     }
 
-    virtual void fft(const float* data, float* re, float* im) override
+    virtual void fft(const double* data, double* re, double* im) override
     {
       const size_t size2 = _size / 2;
       DSPSplitComplex splitComplex;
@@ -995,7 +995,7 @@ namespace audiofft
       splitComplex.imagp = im;
       vDSP_ctoz(reinterpret_cast<const COMPLEX*>(data), 2, &splitComplex, 1, size2);
       vDSP_fft_zrip(_fftSetup, &splitComplex, 1, _powerOf2, FFT_FORWARD);
-      const float factor = 0.5f;
+      const double factor = 0.5f;
       vDSP_vsmul(re, 1, &factor, re, 1, size2);
       vDSP_vsmul(im, 1, &factor, im, 1, size2);
       re[size2] = im[0];
@@ -1003,18 +1003,18 @@ namespace audiofft
       im[size2] = 0.0f;
     }
 
-    virtual void ifft(float* data, const float* re, const float* im) override
+    virtual void ifft(double* data, const double* re, const double* im) override
     {
       const size_t size2 = _size / 2;
-      ::memcpy(_re.data(), re, size2 * sizeof(float));
-      ::memcpy(_im.data(), im, size2 * sizeof(float));
+      ::memcpy(_re.data(), re, size2 * sizeof(double));
+      ::memcpy(_im.data(), im, size2 * sizeof(double));
       _im[0] = re[size2];
       DSPSplitComplex splitComplex;
       splitComplex.realp = _re.data();
       splitComplex.imagp = _im.data();
       vDSP_fft_zrip(_fftSetup, &splitComplex, 1, _powerOf2, FFT_INVERSE);
       vDSP_ztoc(&splitComplex, 1, reinterpret_cast<COMPLEX*>(data), 2, size2);
-      const float factor = 1.0f / static_cast<float>(_size);
+      const double factor = 1.0f / static_cast<double>(_size);
       vDSP_vsmul(data, 1, &factor, data, 1, _size);
     }
 
@@ -1022,8 +1022,8 @@ namespace audiofft
     size_t _size;
     size_t _powerOf2;
     FFTSetup _fftSetup;
-    std::vector<float> _re;
-    std::vector<float> _im;
+    std::vector<double> _re;
+    std::vector<double> _im;
   };
 
 
@@ -1108,9 +1108,9 @@ namespace audiofft
           _size = size;
           _complexSize = AudioFFT::ComplexSize(_size);
           const size_t complexSize = AudioFFT::ComplexSize(_size);
-          _data = reinterpret_cast<float*>(fftwf_malloc(_size * sizeof(float)));
-          _re = reinterpret_cast<float*>(fftwf_malloc(complexSize * sizeof(float)));
-          _im = reinterpret_cast<float*>(fftwf_malloc(complexSize * sizeof(float)));
+          _data = reinterpret_cast<double*>(fftwf_malloc(_size * sizeof(double)));
+          _re = reinterpret_cast<double*>(fftwf_malloc(complexSize * sizeof(double)));
+          _im = reinterpret_cast<double*>(fftwf_malloc(complexSize * sizeof(double)));
 
           fftw_iodim dim;
           dim.n = static_cast<int>(size);
@@ -1122,20 +1122,20 @@ namespace audiofft
       }
     }
 
-    virtual void fft(const float* data, float* re, float* im) override
+    virtual void fft(const double* data, double* re, double* im) override
     {
-      ::memcpy(_data, data, _size * sizeof(float));
+      ::memcpy(_data, data, _size * sizeof(double));
       fftwf_execute_split_dft_r2c(_planForward, _data, _re, _im);
-      ::memcpy(re, _re, _complexSize * sizeof(float));
-      ::memcpy(im, _im, _complexSize * sizeof(float));
+      ::memcpy(re, _re, _complexSize * sizeof(double));
+      ::memcpy(im, _im, _complexSize * sizeof(double));
     }
 
-    virtual void ifft(float* data, const float* re, const float* im) override
+    virtual void ifft(double* data, const double* re, const double* im) override
     {
-      ::memcpy(_re, re, _complexSize * sizeof(float));
-      ::memcpy(_im, im, _complexSize * sizeof(float));
+      ::memcpy(_re, re, _complexSize * sizeof(double));
+      ::memcpy(_im, im, _complexSize * sizeof(double));
       fftwf_execute_split_dft_c2r(_planBackward, _re, _im, _data);
-      detail::ScaleBuffer(data, _data, 1.0f / static_cast<float>(_size), _size);
+      detail::ScaleBuffer(data, _data, 1.0f / static_cast<double>(_size), _size);
     }
 
   private:
@@ -1143,9 +1143,9 @@ namespace audiofft
     size_t _complexSize;
     fftwf_plan _planForward;
     fftwf_plan _planBackward;
-    float* _data;
-    float* _re;
-    float* _im;
+    double* _data;
+    double* _re;
+    double* _im;
   };
 
 
@@ -1180,13 +1180,13 @@ namespace audiofft
   }
 
 
-  void AudioFFT::fft(const float* data, float* re, float* im)
+  void AudioFFT::fft(const double* data, double* re, double* im)
   {
     _impl->fft(data, re, im);
   }
 
 
-  void AudioFFT::ifft(float* data, const float* re, const float* im)
+  void AudioFFT::ifft(double* data, const double* re, const double* im)
   {
     _impl->ifft(data, re, im);
   }
